# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.APIResponse:
    description: Standard API response format with metadata and context information
    properties:
      data:
        description: Response data (varies by endpoint)
      error:
        allOf:
        - $ref: '#/definitions/dto.ErrorResponse'
        description: Error details (only present when success=false)
      meta:
        allOf:
        - $ref: '#/definitions/dto.ResponseMetadata'
        description: Response metadata with request tracking
      pagination:
        allOf:
        - $ref: '#/definitions/dto.PaginationMetadata'
        description: Pagination metadata (for paginated responses)
      success:
        description: Indicates if the request was successful
        example: true
        type: boolean
    type: object
...
  dto.ErrorResponse:
    description: Error response structure with detailed error information
    properties:
      code:
        description: Error code for programmatic handling
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties: true
        description: Additional error details
        type: object
      message:
        description: Human-readable error message
        example: Invalid input data
        type: string
      request_id:
        description: Request ID for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      timestamp:
        description: Error timestamp
        example: "2025-07-24T20:43:16.123456789+07:00"
        type: string
    type: object
...
  dto.PaginationMetadata:
    description: Pagination metadata for paginated responses
    properties:
      has_next:
        description: Whether there is a next page
        example: true
        type: boolean
      has_prev:
        description: Whether there is a previous page
        example: false
        type: boolean
      limit:
        description: Number of items per page
        example: 10
        type: integer
      page:
        description: Current page number (1-based)
        example: 1
        type: integer
      request_id:
        description: Request ID for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      total:
        description: Total number of items
        example: 100
        type: integer
      total_pages:
        description: Total number of pages
        example: 10
        type: integer
    type: object
  dto.ResponseMetadata:
    description: Response metadata containing request tracking and performance information
    properties:
      processing_time_ms:
        description: Request processing time in nanoseconds
        example: 15000000
        type: integer
      request_id:
        description: Unique request identifier for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      timestamp:
        description: Response generation timestamp
        example: "2025-07-24T20:43:16.123456789+07:00"
        type: string
      version:
        description: API version
        example: 1.0.0
        type: string
    type: object
...
  model.Tags:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:4300
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API untuk aplikasi blog Develapar dengan fitur lengkap untuk manajemen
    artikel, komentar, kategori, tag, bookmark, dan like. API menggunakan standard
    response format dengan metadata, request tracking, rate limiting, dan comprehensive
    error handling.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Develapar API
  version: "1.0"
paths:
...
  /tags:
    get:
      description: Get a list of all tags
      produces:
      - application/json
      responses:
        "200":
          description: List of tags
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                    tags:
                      items:
                        $ref: '#/definitions/model.Tags'
                      type: array
                  type: object
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get all tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      description: Create a new tag with a given name
      parameters:
      - description: Tag creation details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.Tags'
      produces:
      - application/json
      responses:
        "201":
          description: Tag successfully created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Tags'
              type: object
        "400":
          description: Invalid payload
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new tag
      tags:
      - Tags
  /tags/{tag_id}:
    delete:
      description: Delete a tag by ID
      parameters:
      - description: ID of the tag to delete
        in: path
        name: tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tag deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid tag ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "404":
          description: Tag not found
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Delete a tag
      tags:
      - Tags
    get:
      description: Get tag details by its ID
      parameters:
      - description: ID of the tag to retrieve
        in: path
        name: tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tag details
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                    tag:
                      $ref: '#/definitions/model.Tags'
                  type: object
              type: object
        "400":
          description: Invalid tag ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get tag by ID
      tags:
      - Tags
    put:
      consumes:
      - application/json
      description: Update an existing tag by ID
      parameters:
      - description: ID of the tag to update
        in: path
        name: tag_id
        required: true
        type: integer
      - description: Tag update details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.Tags'
      produces:
      - application/json
      responses:
        "200":
          description: Tag updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                    tag:
                      $ref: '#/definitions/model.Tags'
                  type: object
              type: object
        "400":
          description: Invalid tag ID or payload
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "404":
          description: Tag not found
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update a tag
      tags:
      - Tags
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"